Projects/KB Azure Knowledgebase/app/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
gunicorn==22.0.0
sqlalchemy==2.0.36
alembic==1.13.1
jinja2==3.1.4
python-multipart==0.0.9
itsdangerous==2.2.0
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
psycopg2-binary==2.9.10
azure-storage-blob==12.19.0
azure-identity==1.15.0
aiofiles==23.2.1
httpx==0.27.0
